hotfix getFwoAlertLogger
